requirements.txt
easygui
trolly
tqdm
awesome-slugify
markdown2
weasyprint
